pull TX table out.
